+++ .android/app/src/main/java/com/ahmed_beheiri/dart_common/host/MainActivity.java
@@ -1,0 +1,6 @@
+package com.ahmed_beheiri.dart_common.host;
+
+import io.flutter.embedding.android.FlutterActivity;
+
+public class MainActivity extends FlutterActivity {
+}